requirements.txt
awesome-slugify==1.6.5
Jinja2==2.11.2
MarkupSafe==1.1.1
regex==2020.11.13
Unidecode==0.4.21
